Split up EventItem component file
And add http_response event
