+++ packages/agents-ui/src/components/AgentDetails/EventsView/EventItem/index.ts
@@ -1,0 +1,1 @@
+export { EventItem } from "./EventItem";
