+++ packages/agents-ui/src/components/AgentDetails/EventsView/EventsCategoriesFilter.tsx
@@ -3,6 +3,7 @@ import {
   FpDropdownMenuCheckboxItem,
   FpDropdownMenuContent,
   FpDropdownMenuItem,
+  FpDropdownMenuLabel,
   FpDropdownMenuPortal,
   FpDropdownMenuSeparator,
   FpDropdownMenuTrigger,
@@ -11,6 +12,7 @@ import { cn } from "@/lib/utils";
 import { AllEventCategories, usePlaygroundStore } from "@/store";
 import type { AgentInstanceParameters } from "@/types";
 import { CaretSortIcon } from "@radix-ui/react-icons";
+import { ListFilter } from "lucide-react";
 
 export function EventCategoriesFilter(props: AgentInstanceParameters) {
   const selectedCategories = usePlaygroundStore(
@@ -31,7 +33,7 @@ export function EventCategoriesFilter(props: AgentInstanceParameters) {
 
   return (
     <div className="flex items-center gap-2">
-      Filter:
+      <ListFilter className="w-3.5 h-3.5" />Filter
       <FpDropdownMenu>
         <FpDropdownMenuTrigger
           className={cn(
@@ -43,15 +45,26 @@ export function EventCategoriesFilter(props: AgentInstanceParameters) {
             "data-[state=open]:bg-muted",
             "rounded-sm",
             "group/dropdown",
+            "py-1"
           )}
         >
           <div className="grow-1 w-full text-start text-sm text-muted-foreground px-2">
-            {"Event categories"}
+            {selectedCategories.length === AllEventCategories.length ? (
+              "All categories"
+            ) : (
+              <>
+                {selectedCategories.length}{" "}
+                {selectedCategories.length === 1 ? "category" : "categories"}
+              </>
+            )}
           </div>
           <CaretSortIcon className="w-3 h-3 mr-1 flex-shrink-0 opacity-0 group-hover:opacity-100 group-focus:opacity-100 group-data-[state=open]/dropdown:opacity-100" />
         </FpDropdownMenuTrigger>
         <FpDropdownMenuPortal>
           <FpDropdownMenuContent align="start">
+            <FpDropdownMenuLabel className="text-muted-foreground uppercase pt-2">
+              Categories
+            </FpDropdownMenuLabel>
             {AllEventCategories.map((category) => (
               <FpDropdownMenuCheckboxItem
                 checked={selectedCategories.includes(category)}
@@ -62,11 +75,14 @@ export function EventCategoriesFilter(props: AgentInstanceParameters) {
               </FpDropdownMenuCheckboxItem>
             ))}
             <FpDropdownMenuSeparator />
-            <FpDropdownMenuItem onChange={() => resetEventCategories()}>
-              Default events
+            <FpDropdownMenuLabel className="text-muted-foreground uppercase  pt-2">
+              Actions:
+            </FpDropdownMenuLabel>
+            <FpDropdownMenuItem onClick={() => resetEventCategories()}>
+              Select default categories
             </FpDropdownMenuItem>
-            <FpDropdownMenuItem onChange={() => unselectAllEventCategories()}>
-              Hide all events
+            <FpDropdownMenuItem onClick={unselectAllEventCategories}>
+              Deselect all
             </FpDropdownMenuItem>
           </FpDropdownMenuContent>
         </FpDropdownMenuPortal>
